Fix debug elements to use mesh coordinates instead of pixel coordinates
BREAKING CHANGE: Debug element/location coordinates now in mesh space

Previously, WithDebugElement and WithDebugLocation accepted pixel
coordinates (0-1920, 0-1080), but this didn't match the mesh coordinate
system. Debug elements would not align with mesh geometry.

Changed:
- DebugElement: SourceX/Y, TargetX/Y from int to float64
- DebugLocation: X, Y from int to float64
- Coordinates now use mesh space (same as vertices)
- Transform applied automatically during rendering

Example:
  // OLD (pixel coordinates - BROKEN)
  WithDebugElement("edge", 100, 100, 500, 500)

  // NEW (mesh coordinates - CORRECT)
  WithDebugElement("edge", 10.5, 20.3, 100.7, 200.1)

Updated rendering:
- renderDebugElements now applies transform
- renderDebugLocations now applies transform
- Debug elements scale/translate with mesh geometry

Updated tests:
- Use mesh coordinates (0-10 range matching vertices)
- Added vertices to empty mesh test for proper transform
- All tests passing

Your code is correct now! Just remove the int() casts:
  // Before (broken):
  rasterize.WithDebugElement(line.Name,
      int(line.A.X), int(line.A.Y),
      int(line.B.X), int(line.B.Y))

  // After (works):
  rasterize.WithDebugElement(line.Name,
      line.A.X, line.A.Y,
      line.B.X, line.B.Y)

+++ rasterize/config.go
@@ -3,19 +3,21 @@ package rasterize
 import "image/color"
 
 // DebugElement represents a debug line with a label.
+// Coordinates are in mesh space and will be transformed to image coordinates.
 type DebugElement struct {
 	Name    string
-	SourceX int
-	SourceY int
-	TargetX int
-	TargetY int
+	SourceX float64
+	SourceY float64
+	TargetX float64
+	TargetY float64
 }
 
 // DebugLocation represents a debug point location with a label.
+// Coordinates are in mesh space and will be transformed to image coordinates.
 type DebugLocation struct {
 	Name string
-	X    int
-	Y    int
+	X    float64
+	Y    float64
 }
 
 // Config holds options for rasterizing a mesh to an image.
